Top five controller can now switch between acres/animals
Added the new URL to the farmData service, (qualified by year to remove
the apparent duplicates), and switch between them by adding two buttons
to set the current category.

To make the divs transition size nicely, you have to tell angular not
to delete and replace them when you change the underlying data. To do
that I had to get rid of the filters in the ng-repeat, use a track by,
and do the sorting and limiting in the controller after the fetch.